+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+- Fix multi-signature wallet detection
+
 ## 0.2.1 - 2021-01-14
 
 Migrated from [`bwt-dev/bwt`](https://github.com/bwt-dev/bwt) into a standalone repo